refactor: extract node representation
