+++ app/src/main/java/com/neo/speaktouch/controller/SpeechController.kt
@@ -22,12 +22,8 @@ import android.content.Context
 import android.speech.tts.TextToSpeech
 import androidx.core.view.accessibility.AccessibilityNodeInfoCompat
 import com.neo.speaktouch.model.Text
-import com.neo.speaktouch.utils.NodeValidator
 import com.neo.speaktouch.utils.Reader
-import com.neo.speaktouch.utils.extension.getContent
-import com.neo.speaktouch.utils.extension.iterator
-import org.json.JSONArray
-import org.json.JSONObject
+import com.neo.speaktouch.utils.Node
 import timber.log.Timber
 
 class SpeechController(
@@ -66,39 +62,5 @@ class SpeechController(
 
     fun shutdown() {
         textToSpeech.shutdown()
-    }
-}
-
-data class Node(
-    val nodeInfo: AccessibilityNodeInfoCompat,
-    val content: CharSequence? = nodeInfo.getContent(),
-    val children: List<Node> = nodeInfo.iterator().map { Node(it) }.toList()
-) {
-    fun geJson(): JSONObject {
-
-        val json = JSONObject()
-
-        json.put("content", content)
-        json.put("className", nodeInfo.className)
-        json.put("isReadableAsChild", NodeValidator.isReadableAsChild(nodeInfo))
-
-        val childrenJson = JSONArray()
-
-        children.forEachIndexed { index, node ->
-            childrenJson.put(index, node.geJson())
-        }
-
-        json.put("children", childrenJson)
-
-        return json
-    }
-}
-
-private fun <T, R> Iterator<T>.map(function: (T) -> R): List<R> {
-
-    return buildList {
-        while (hasNext()) {
-            add(function(next()))
-        }
     }
 }
